Use pytest for testing
